+++ src/components/Modals/Onboarding/OnboardingTransfer.vue
@@ -49,7 +49,7 @@
 
 <script lang="ts">
 import { PropType, defineComponent } from 'vue';
-import { examData as reqsData, ExamRequirements } from '@/requirements/data/exams/ExamCredit';
+import { examSubjects, getExamCredit } from '@/requirements/requirement-exam-utils';
 import OnboardingTransferSwimming from './OnboardingTransferSwimming.vue';
 import OnboardingTransferCreditsSource from './OnboardingTransferCreditsSource.vue';
 
@@ -63,8 +63,6 @@ type Data = {
   tookSwimTest: 'yes' | 'no';
   scoresAP: readonly number[];
   scoresIB: readonly number[];
-  subjectsAP: readonly string[];
-  subjectsIB: readonly string[];
   placeholderText: string;
   examsAP: FirestoreAPIBExam[];
   examsIB: FirestoreAPIBExam[];
@@ -73,36 +71,6 @@ type Data = {
 
 const scoresAP = [5, 4, 3, 2, 1];
 const scoresIB = [7, 6, 5, 4, 3, 2, 1];
-const existingAP: Record<string, boolean> = {};
-const unmodifiedReqsData = { ...reqsData };
-// filter duplicate exam names and ones already selected
-reqsData.AP = reqsData.AP.filter(ap => {
-  const inExisting = ap.name in existingAP;
-  existingAP[ap.name] = true;
-  return !inExisting;
-});
-const existingIB: Record<string, boolean> = {};
-// filter duplicate exam names and ones already selected
-reqsData.IB = reqsData.IB.filter(ib => {
-  const inExisting = ib.name in existingIB;
-  existingIB[ib.name] = true;
-  return !inExisting;
-});
-const subjectsAP = reqsData.AP.map(it => it.name);
-const subjectsIB = reqsData.IB.map(it => it.name);
-
-export const getExamCredit = (exam: FirestoreAPIBExam): number => {
-  const allExamsWithSameName: ExamRequirements[] = unmodifiedReqsData[exam.type].filter(
-    it => it.name === exam.subject
-  );
-  let mostPossibleCredit = 0;
-  for (const examWithSameName of allExamsWithSameName) {
-    if (exam.score >= examWithSameName.fulfillment.minimumScore) {
-      mostPossibleCredit = Math.max(mostPossibleCredit, examWithSameName.fulfillment.credits);
-    }
-  }
-  return mostPossibleCredit;
-};
 
 export default defineComponent({
   components: {
@@ -131,8 +99,6 @@ export default defineComponent({
         typeof this.onboardingData.tookSwim !== 'undefined' ? this.onboardingData.tookSwim : 'no',
       scoresAP,
       scoresIB,
-      subjectsAP,
-      subjectsIB,
       examsAP,
       examsIB,
       classes: transferClasses,
@@ -143,16 +109,23 @@ export default defineComponent({
     totalCredits(): number {
       let count = 0;
       [...this.examsAP, ...this.examsIB].forEach(exam => {
-        count += this.getExamCredit(exam);
+        count += getExamCredit(exam);
       });
       this.classes.forEach(clas => {
         count += clas.credits;
       });
       return count;
     },
+    subjectsAP(): string[] {
+      const currentSubjects = new Set(this.examsAP.map(exam => exam.subject));
+      return examSubjects.AP.filter(subject => !currentSubjects.has(subject));
+    },
+    subjectsIB(): string[] {
+      const currentSubjects = new Set(this.examsIB.map(exam => exam.subject));
+      return examSubjects.IB.filter(subject => !currentSubjects.has(subject));
+    }
   },
   methods: {
-    getExamCredit,
     getTransferClassSearchboxPlaceholder(text: string): string {
       return text !== this.placeholderText ? text : '"CS1110", "Multivariable Calculus", etc';
     },
