+++ src/requirements/requirement-frontend-computation.ts
@@ -1,5 +1,5 @@
 import { SWIM_TEST_COURSE_ID } from './data/constants';
-import getCourseEquivalentsFromUserExams from './data/exams/ExamCredit';
+import getCourseEquivalentsFromUserExams from './requirement-exam-utils';
 import {
   courseIsAPIB,
   convertFirestoreSemesterCourseToCourseTaken,
